feat: split pages in their own folders
