+++ src/Modules/Guilds/pages/AllProposals/AllProposals.styled.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import { Box } from 'Components/Primitives/Layout';
+
+export const ProposalsList = styled(Box)`
+  margin-top: 1rem;
+`;
+
+export const ProposalListWrapper = styled.div<{ isSearchOpened: boolean }>`
+  height: ${({ isSearchOpened }) => (isSearchOpened ? '43vh' : '50vh')};
+  @media only screen and (min-width: 768px) {
+    height: ${({ isSearchOpened }) => (isSearchOpened ? '69vh' : '75vh')};
+  }
+`;
