+++ src/Modules/Guilds/pages/AllProposals/index.ts
@@ -1,0 +1,1 @@
+export { default as AllProposals } from './AllProposals';
