+++ src/Modules/Guilds/pages/Governance/index.ts
@@ -1,0 +1,1 @@
+export { default as Governance } from './Governance';
