+++ src/Modules/Guilds/pages/Guilds/Guilds.styled.tsx
@@ -1,0 +1,24 @@
+import styled from 'styled-components';
+import { Box } from 'Components/Primitives/Layout';
+
+export const PageContainer = styled(Box)`
+  display: grid;
+  grid-template-columns: 1fr;
+
+  /* Medium devices (landscape tablets, 768px and up) */
+  @media only screen and (min-width: 768px) {
+    grid-template-columns: 300px minmax(0, 1fr);
+  }
+`;
+
+export const SidebarContent = styled(Box)`
+  @media only screen and (min-width: 768px) {
+    margin-right: 0.5rem;
+  }
+`;
+
+export const PageContent = styled(Box)`
+  @media only screen and (min-width: 768px) {
+    margin-left: 0.5rem;
+  }
+`;
