+++ src/Modules/Guilds/pages/Guilds/index.tsx
@@ -1,0 +1,1 @@
+export { default as GuildsPage } from './Guilds';
