+++ src/Modules/Guilds/pages/LandingPage/index.ts
@@ -1,0 +1,1 @@
+export { default as LandingPage } from './LandingPage';
