+++ src/Modules/Guilds/pages/Proposal/index.ts
@@ -1,0 +1,1 @@
+export { default as ProposalPage } from './Proposal';
